+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- `LaunchToRendezvous()` and `LaunchToTargetPlane()` methods in AscentAutopilot ([#8](https://github.com/Genhis/KRPC.MechJeb/issues/8))
 
 ## [0.5.0] - 2019-06-11
 ### Added